Tests for empty dmatrix.
